+++ src/radiator/config.clj
@@ -1,3 +1,4 @@
+
 (ns radiator.config
   (:require [clojure.spec.alpha :as spec]
             [clojure.spec.alpha :as s]))
@@ -18,11 +19,12 @@
 (s/def ::name ::non-empty-string)
 (s/def ::uri #(uri? (java.net.URI. %)))
 (s/def ::api-key ::non-empty-string)
-
+(s/def ::endpoint (s/keys :opt-un [::name ::uri]))
+(s/def ::endpoints (s/coll-of ::endpoint))
 (s/def ::aws (s/keys :req-un [::uri ::api-key]))
 (s/def ::gitlab-pipeline (s/keys :req-un [::name ::uri ::api-key]))
 (s/def ::gitlab-pipelines (s/coll-of ::gitlab-pipeline))
-(s/def ::project (s/keys :opt-un [::aws ::gitlab-pipelines]))
+(s/def ::project (s/keys :opt-un [::aws ::gitlab-pipelines ::endpoints]))
 (s/def ::projects (s/coll-of ::project))
 
 (def projects